uievents/order-of-events/mouse-events/click-order-manual.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>Click-related events should fire in the correct order</title>
    <link rel="author" title="Sandra Sun">
    <link rel="help" href="https://w3c.github.io/uievents/#events-clickevent-event-order">
    <meta name="flags" content="interact">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <ol>
      <li>Double-click into the blue area.</li>
      <li>Click the "Done" button.</li>
    </ol>
    <div id="blue" style="width:400px; height:300px; background-color: blue;"></div>
    <button type="button" id="done">Done</button>
</body>
<script>
setup({explicit_timeout: true});
var done = false;
var started = false;
var events = [];
var targets = [];

function record(e) {
  if (e.type == 'mousedown') {
    started = true;
  }
  if (!started || done) {
    return;
  }
  events.push(e.type);
  targets.push(e.target.id);
}

var relevantEvents = [
  'mousedown',
  'mouseup',
  'mousemove',
  'click',
  'dblclick'
];

function finish() {
  done = true;
}

function checkOrder(event_types) {
  var order = ['initial', 'mousedown', 'mousemove', 'mouseup', 'click',
    'mousemove', 'mousedown', 'mousemove', 'mouseup', 'click', 'dblclick',
    'mousemove'];
  var optional = [false, false, true, false, false,
    true, false, true, false, false, false, true];
  var state = 0;

  for (var i = 0; i < event_types.length; ++i) {
    // Move to the next state.
    if (state + 1 < order.length && event_types[i] == order[state + 1]) {
      state = state + 1;
      continue;
    }
    // Stay at the current state.
    if (optional[state] && event_types[i] == order[state]) {
        continue;
    }
    // Jump over the next state.
    if (state + 2 < order.length && optional[state + 1]
      && event_types[i] == order[state + 2]) {
      state += 2;
      continue;
    }
    return false;
  }

  return (state >= order.length - 2);
}

function checkTarget(event_targets) {
  for (var i = 0; i < event_targets.length; ++i) {
    if (event_targets[i] != 'blue')
      return false;
  }
  return true;
}

window.onload = function () {
  var blue = document.getElementById('blue');

  for (var i = 0; i < relevantEvents.length; i++) {
    blue.addEventListener(relevantEvents[i], record, false);
  }

  async_test(function(t) {
    document.getElementById('done').addEventListener('click', function () {
      finish();
      t.step(function () {
        assert_true(
          checkOrder(events),
          'Click-related events should fire in the correct order'
        );
        assert_true(
          checkTarget(targets),
          'Click-related events should fire at the correct targets'
        );
        t.done();
      });
    }, false);
  }, 'Click-related events should fire in the correct order');
};
    </script>
</html>
